Run test cases
Add logic for running individual test cases. Now that we've compiled the
constraints, all we need to do is Review() the specified objects and
compare the Results with what the user specified.

This PR does not add Assertion functionality - we'll do that for alpha,
not pre-alpha.

Also add unit tests that verify we are, in fact, running the test cases.

Signed-off-by: Will Beason <willbeason@google.com>