Make Context usage consistent
In trying to upgrade our controller-runtime dependency to v0.9.0, I
found that the upgrade introduced several bugs. controller-runtime is
now being more thorough in its usage of Context, which meant we had
cases where we _had_ a context, but weren't properly passing it down the
call chain.

This CL eliminates all usage of context.Background() from pkg/, which
does resolve that part of the upgrade issue.

The other issue (to be resolved separately) is that we both depend on
the controller-runtime package in vendor/ and a different version in
third_party/.

Signed-off-by: Will Beason <willbeason@google.com>